requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
jupyter
notebook
pyarrow
datasets
python-dotenv
streamlit
plotly